Day 5 Solution
I cannot for the life of me figure out why trying to combine the two bound functions into one, which also takes a character signifying if it wants a lower or upper bound, results in not getting an answer. I think its in a loop and cant see where. Weird. Works with seperate functions tho, like here